+++ src/RcsbChartWeb/RcsbChartComponent/VictoryChartImplementations/Components/TooltipFactory.tsx
@@ -3,10 +3,12 @@ import * as React from "react";
 import {VictoryTooltipProps} from "victory-tooltip";
 import {ChartTools} from "../../../RcsbChartTools/ChartTools";
 import {ChartDataInterface} from "../../../RcsbChartDataProvider/ChartDataProviderInterface";
+import {ChartConfigInterface} from "../../ChartConfigInterface";
 
-export class TooltipFactory extends React.Component<VictoryTooltipProps, any> {
+type TooltipPropsType = VictoryTooltipProps & {tooltipText?:ChartConfigInterface["tooltipText"]};
+export class TooltipFactory extends React.Component<TooltipPropsType, any> {
 
-    public static getTooltip(props:VictoryTooltipProps) {
+    public static getTooltip(props:TooltipPropsType) {
         return (<VictoryTooltip
             {...props}
             text={TooltipFactory.text}
@@ -19,8 +21,10 @@ export class TooltipFactory extends React.Component<VictoryTooltipProps, any> {
         />);
     }
 
-    private static text(props:VictoryTooltipProps):string {
+    private static text(props:TooltipPropsType):string {
         const d: ChartDataInterface = (props.datum as ChartDataInterface);
+        if(typeof props.tooltipText === "function")
+            return props.tooltipText(d);
         return d.y.toString() + (d.yc > 0? (" of " + (d.y+d.yc).toString()) : "") + " group members\n" +
             "Click to refine group\n" +
             "Shift-click to search";