+++ src/RcsbFvExamples/ResidueDistribution.ts
@@ -1,3 +1,3 @@
 import {buildResidueDistribution} from "../RcsbGroupWeb/RcsbGroupBuilder";
 
-buildResidueDistribution("pfv", "instance", "AF_AFA0A023GPK8F1.A");
+buildResidueDistribution("pfv", "entry", "MA_MABAKCEPC0005", ["MA_QA_METRIC_LOCAL_TYPE_PLDDT", "MA_QA_METRIC_LOCAL_TYPE_OTHER"]);