+++ src/RcsbGroupWeb/RcsbGroupView/RcsbResidueChart/RcsbResidueChartComponent.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import {ChartMapType, GroupChartLayout} from "../RcsbGroupChart/GroupChartLayout";
-import {ResidueChartInterface, ResidueChartTools as RCT} from "./ResidueChartTools";
+import {ResidueChartInterface, ResidueChartTools as RCT} from "./ResidueChartTools/ResidueChartTools";
 import classes from "../RcsbGroupMembers/Components/scss/group-display.module.scss";
 import {Container} from "react-bootstrap";
 import {RcsbChartInterface} from "../../../RcsbSeacrh/FacetTools";
@@ -10,7 +10,7 @@ interface RcsbResidueChartState {
     chartMap: ChartMapType;
 }
 
-export class RcsbResidueChartComponent extends React.Component <ResidueChartInterface, RcsbResidueChartState> {
+export class RcsbResidueChartComponent extends React.Component <ResidueChartInterface & {facetLayoutGrid?:string[];}, RcsbResidueChartState> {
 
     render(): JSX.Element {
         if (this.state?.layout?.flat().filter((e) => (this.state?.chartMap?.get(e)))) {
@@ -31,7 +31,7 @@ export class RcsbResidueChartComponent extends React.Component <ResidueChartInte
     }
 
     private async updateState(): Promise<void> {
-        const charts: RcsbChartInterface[] = await RCT.getResidueDistribution(this.props);
+        const charts: RcsbChartInterface[] = (await (await RCT.getResidueDistribution(this.props))).filter(chart=>(!this.props.facetLayoutGrid || this.props.facetLayoutGrid.includes(chart.attribute)));
         this.setState({
             layout: charts.map(c=>c.attribute),
             chartMap: charts.reduce<ChartMapType>((prev,curr)=>(prev.set(curr.attribute,{chart: curr})), new Map())