+++ src/RcsbGroupWeb/RcsbGroupView/RcsbResidueChart/ResidueChartFactories/ResidueDistributionFactoryImpl/TrackDistributionFactory.ts
@@ -9,7 +9,7 @@ import {
 import {MultipleTrackDistributionFactory} from "./MultipleTrackDistributionFactory";
 import {NumericalTrackDistributionFactory} from "./NumericalTrackDistributionFactory";
 
-export class TrackDistributionFactory implements ResidueDistributionFactoryInterface<[string]> {
+export class TrackDistributionFactory implements ResidueDistributionFactoryInterface<[string,number]> {
 
     private readonly distributionConfig: RcsbDistributionConfig;
     private readonly multipleTrackDistributionFactory: MultipleTrackDistributionFactory
@@ -22,13 +22,13 @@ export class TrackDistributionFactory implements ResidueDistributionFactoryInter
 
     }
 
-    getDistribution(tracks:TrackManagerInterface[],blockType:string): ResidueDistributionInterface {
+    getDistribution(tracks:TrackManagerInterface[], blockType:string, numberResidues: number): ResidueDistributionInterface {
 
         switch(this.distributionConfig.getBlockConfig(blockType).contentType) {
             case "binary":
-                return this.multipleTrackDistributionFactory.getDistribution(tracks, blockType);
+                return this.multipleTrackDistributionFactory.getDistribution(tracks, blockType, numberResidues);
             case "numerical":
-                return this.numericalTrackDistributionFactory.getDistribution([tracks[0]], blockType);
+                return this.numericalTrackDistributionFactory.getDistribution([tracks[0]], blockType, numberResidues);
         }
 
     }